Remove tests that fails in certain envs
This test was deleted because it was not passing in concourse
which runs mysql v5.5. We did see that the size constraint was
honored on local workstations running mysql v5.7.

The controller which populates this column has validation on the
size of the context_id, so we feel safe deleting this lower level test.

[#134702861](https://www.pivotaltracker.com/story/show/134702861)

Signed-off-by: Tyler Schultz <tschultz@pivotal.io>