Update drain int tests to accomodate new parameters
passed from the agent. The agent no longer sends the job_unchanged
as a parameter. This was sent when the agent saw that the old spec sha1
and the new spec sha1 were equal. This case is problematic now that the
sha1s are pointerized. This case was deemed not worth the effort to
keep, so the case was removed.

[#134619633](https://www.pivotaltracker.com/story/show/134619633)

Signed-off-by: Tyler Schultz <tschultz@pivotal.io>