Use the digest from the uploaded manifest
- Stop recalculating the sha1 on uploaded asset.
- In the case a compiled release was uploaded, and did not have a source
release, then the source tarball for the same release was uploaded,
the digest of the source release was overriding the digest of the compiled
release, and therefore the compiled release was using the wrong digest.
We suspect this problem lead to the (now removed) code that shasummed the uploaded asset.

[#139399691](https://www.pivotaltracker.com/story/show/139399691)

Signed-off-by: Adrian Zankich <azankich@pivotal.io>