Fix broken migration
A migration, originally released in 257.x was updated in 258.x. The
update added an index to the local_dns_blobs table. Directors that have
been upgraded from 257 will not have this index. Therefore, the
update_sha1_column_sizes migration breaks when it tries to drop the
missing index. Furthermore we have deeemed the we do not need the index.
Hence we have decided to drop adding it back. A subsequent migration
will remove the index if it exists. This should address directors that
successfully ran the update_sha1_column_sizes migration before this fix.

Cherry-picked from 42cf699

[#140068815](https://www.pivotaltracker.com/story/show/140068815)

Signed-off-by: Dennis Leon <dleon@pivotal.io>