Introduce concourse task for GcscliBlobstoreClient
The task is based off of the pre-existing S3 task.

The blob.blobstore_client Dockerfile is updated to include the
gcloud SDK. This is necessary as functional tests require the
existence of buckets with specific permissions. locales is
also installed as it is not available by default on the
current ubuntu:latest.