Serialize template rendering
This is being done because some release templates declare duplicate consts.
These constants were used and interpolated later in the template. When
rendering multiple templates in parallel, the const value would change
as it was reassigned by the different templates, resulting in unexpected
content in the rendered template. This is a short term solution while we
put together a more comprehensive solution.

[#148024849](https://www.pivotaltracker.com/story/show/148024849)

Signed-off-by: Arthur Chan <achan@pivotal.io>