Update director configuration to include agent fields
The agent configuration was previously omitted, this corrects that

It was difficult to determine what was the failing for the agent
in end to end tests given lack of a stemcell including bosh-gcscli.
Manually including bosh-gcscli in the stemcell enabled useful
results from the end to end test, catching this fault.