Available zone ID for address encoding.
This adds a table that records the ids generated to encode the AZs. We
don't want to have to track the types of the AZs i.e. in the dns_records
so it was decided to create a separate table for them. Recording the ids
we used is necessary so that the ids are consistent and deterministic
across deploys.

[#136925949](https://www.pivotaltracker.com/story/show/136925949)

Signed-off-by: Arthur Chan <achan@pivotal.io>