As an operator, I expect to be able to retrieve exact same manifest that was submitted.
Manifest text, which is sent to director, is stored at database as raw_manifest and retrieved from it without modifications.

Several variables were renamed to show whether they are text or hash.
New :manifest_file option was added to deploy method of IntegrationExampleGroup to support ability to pass manifest files from fs directly.
Usage of manifest hash which is based on manifest text in the actual deployment process is out of the scope

(#150869116)[https://www.pivotaltracker.com/story/show/150869116]

Signed-off-by: Konstantin Maksimov <kmaksimov@ru.ibm.com>