Remove caches from DnsEncoder in favor of DB
DnsEncoder is instantiated often. Creating it once and passing it along,
via InstancePlan, does not work, as it needs to get updated, i.e.
after instance creation.

Therefore we remove two of the caches, 'Instance uuid to name' and the one
for 'network name to id'. Thus we avoid two full table scans on every
instantiation.

[#153991487](https://www.pivotaltracker.com/story/show/153991487)