Fix stemcell disassociation
This change updates the `cleanup_stemcell_references_stage` to not
modify the array of deployment stemcells while we are iterating over it.
Previously, this would cause issues as the ordering was unstable if we
removed an element.

This change also updates the unit tests object names and values for
clarity.

[#155730396](https://www.pivotaltracker.com/story/show/155730396)

Signed-off-by: James Myers <jfmyers9@gmail.com>