Orphan `create-swap-delete` VMs on deletion
This change will unmount persistent disks, detach persistent disks, and
orphan VMs on deletion if using the `create-swap-delete` VM update
strategy. If the agent is unresponsive, the VM will be deleted
synchronously. This change takes effect when deleting an instance group,
scaling down instances, or deleting a deployment.

**Note**: This change does not delete VMs synchronously that have opted
into using static ips. This will be implemented in a future commit.

[#154011728](https://www.pivotaltracker.com/story/show/154011728)