Orphan extra inactive VMs when updating an instance
This change will orphan inactive VMs that are not scheduled to become
the active VM. This ensures that these VMs will eventually be cleaned up
by the background cleanup process.

This change also modifies the logic for releasing obsolete network
plans. Previously, we used to release all obsolete network plans when a
VM was deleted. Now when deleting a VM due to an unresponsive agent, we
will only attempt to release the obsolete network plans for ip addresses
that are related to the VM being deleted. This prevents invalid re-use
of a in-use IP address when "create-swap-delete"ing. This change also
removes a call to `release_obsolete_network_plans` as this logic was
duplicated in the VmCreator.

This change also adds a comprehensive integration test for orphaning VMs
throughout multiple deploys with various end results.

[#157379394](https://www.pivotaltracker.com/story/show/157379394)

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>