Remove network plans associated with orphaned vms
This change re-adds the logic to "release" all obsolete network plans
while updating an instance. Before releasing obsolete network plans, we
need to untrack the network plans that are associated with orphaned VMs,
as the VMs still exist, and their IPs are still in use.

This was necessary due to the following case for create-swap-delete:
- Deploy failure for instance with static ip when creating VM
- Remove static IP and redeploy
The above scenario resulted in a network plan with a static ip
reservation that had no corresponding VM. We were previously never
releasing this IP address.

[finishes #157379394](https://www.pivotaltracker.com/story/show/157379394)
[finishes #157580640](https://www.pivotaltracker.com/story/show/157580640)

Co-authored-by: Danny Berger <dberger@pivotal.io>