Revert compilation-double-deletion fix
- the compilation-double-deletion changes caused unit test failures that
turned out to be challenging to fix quickly

- Specifically, the instance_deleter unit test stubbed its dependencies
such that its persistent disks are not persisted to the database, which
results in a change in the count of bosh events when vm deleter attempts
to clean up.  Naive attempts to persist the data resulted in a foreign
key constraint violation, so this requires more careful work to fix.

This reverts commit 5b8cb57a1b4ca64f820b5868b7759c2128408dbf.
This reverts commit d4f8be80824620182550844071afbe9f7f93c87f.