Always delete compilation instance
This change updates the `destroy_instance` method to orphan VMs when
necessary and then still call the instance_deleter. Because the VMs are
already orphaned, the vm_deleter should not take any action as there
will be no VMs associated with the instance. It also releases the
network plans after orphaning so that the IP addresses are not
disassociated with the orphaned vms.

This change also improves upon the bug fix introduced in story #158130962
opting instead to refresh stale ORM models. This ensures that the call
to the vm_deleter inside the instance_deleter will not attempt to delete
VMs that have already been removed.

[#158512624](https://www.pivotaltracker.com/story/show/158512624)
[#158130962](https://www.pivotaltracker.com/story/show/158130962)

Co-authored-by: James Myers <jmyers@pivotal.io>