Ensure sandbox db always migrates correctly
Previously we were running into temporal errors when certain tests that
require dns to be disabled run before other tests; this order caused
other tests to fail because certain migrations were not run. This change
ensures that the sandbox will always have a properly migrated db when
the sandbox is reset and that test execution should not matter.

[#159321346](https://www.pivotaltracker.com/story/show/159321346)

Co-authored-by: Chris Dutra <cdutra@pivotal.io>