Only cleanup unused stemcell references on success
This change updates the director to only remove stemcells from a
deployment if the deployment is successful. Previously on failed
deploys, the director would update the deployment to only refer to the
desired stemcell. This resulted in confusing output from the CLI and
also allowed users to delete stemcells that were still in use by various
deployments.

Fixes #1427.

[finishes #159864892](https://www.pivotaltracker.com/story/show/159864892)

Co-authored-by: James Myers <jmyers@pivotal.io>