Use asymmetric key encryption for UAA in integration
We only restart the UAA process when the encryption method is changed.
This was a major source of failure in CI as restarting UAA can often
take longer than 10 minutes. Symmetric key encryption is no longer
recommended by the UAA team, and is not how we currently deploy modern
BOSH directors.

Co-authored-by: Max Petersen <mpetersen@pivotal.io>