Use custom RDS parameter-group to disable GTID
GTID (Global Transaction ID) is a feature that was recently turned on
in AWS RDS by default. This feature results in mysqldump files that
include commands that cannot be run by non-SUPER-users (like the one
that we are given to dump the file). That means that BBR backup works
but BBR restore will fail against the default RDS setup. We've got this
working by turning off GTID with a custom parameter group.

[#161030792](https://www.pivotaltracker.com/story/show/161030792)
[#160976166](https://www.pivotaltracker.com/story/show/160976166)

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>