Consider CPI aliases when creating vm for an existing instance
This change updates the create_vm_step to respect CPI alias names when
finding the stemcell for an existing instance. This code path is
exercised when apply resolutions to VMs with unresponsive agents.

This fixes an issue in which the director would error when resurrecting
failing instances after migrating to a CPI config and AZs that reference
CPIs.

This change also formats the instance_spec to adhere to rubocop.

[finishes #161852277](https://www.pivotaltracker.com/story/show/161852277)

Co-authored-by: James Myers <jmyers@pivotal.io>