Store version in releases in runtime config as string
When saving the runtime config, we must ensure that the version field
of the releases block is a string such that we don't see unnecessary diffs.

Before this commit, we saw an unnecessary diff when:
1. deploy dummy (containing release {name: a, version: 1})
2. update runtime config (containing release {name: a, version: 1})
3. redeploy dummy
because the deployment manifest stores the version as string whereas the
runtime config stores the version type as-is (i.e. as integer if the user
provides an integer).

[#153632014](https://www.pivotaltracker.com/story/show/153632014)

Co-authored-by: Denis Holdeew <denis.holdeew@sap.com>
Co-authored-by: David Ansari <david.ansari@sap.com>