UpdateDeployment should respect configuration
Before this change, no matter what config IDs were being passed into
UpdateDeployment, it was using an empty array for the cloud config
models and runtime config models when preparing the update deployments.
This change adds test coverage for that code path and fixes the loading
of those models at the beginning of UpdateDeployment#perform.