Use consistent names for generating link group IDs
Previously we were conflating the link provider intent's original name
and alias name when generating and storing link group IDs. This change
introduces the canonical name for a link provider intent, which is used
in all contexts to enforce consistency. The canonical name is the alias
name for explicit links and the original name for implicit links.

[#161942444](https://www.pivotaltracker.com/story/show/161942444)

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Josh Aresty <joshua.aresty@emc.com>