Test create-swap-delete when provider unresponsive
When link providers and consumers are in the same deployment and a
create-swap-delete deployment fails with an unresponsive link provider,
we delete the provider VM in parallel with updating its consumer VMs.
If using IP addresses to fill in consumer job templates, this sometimes
results in the unresponsive provider VM's IP address being filled into
consumer templates before the link provider is recreated.  This test
validates that when using DNS addresses to reference link providers, the
correct address is used so the problem that happens when using IP
addresses is not an issue so long as DNS is being used.

[#161611131](https://www.pivotaltracker.com/story/show/161611131)