Ensure deployment includes tags from runtime-config
This fixes a problem where ressurection wouldn't set all the correct
tags on VMs that were recreated. But also affects everything that was
previously using pre-stored tags (such as AttachDisk, ressurection,
etc). All of those actions should now result in resources that contain
tags from both the deployment and the applicable runtime config.

We opted to merge in the tags as part of Models::Deployment since it
would fix several of these other edge cases without needing to duplicate
the merging behavior and some significant refactoring.

[#162933727](https://www.pivotaltracker.com/story/show/162933727)

Co-authored-by: Danny Berger <dberger@pivotal.io>