Remove syslog release as an external dependency from external db tests
This has been flakey as it requires the public internet to download that
asset. If the director successfully deploys we should have confidence
that it was able to reach the database. This is proven by the fact that
the negative cases check that it doesn't deploy successfully.

This change also moves the deletion of the inner bosh to an after each
rather than a defer.

[#163241835](https://www.pivotaltracker.com/story/show/163241835)

Co-authored-by: James Myers <jmyers@pivotal.io>