Refactor compiled package detection into a new object
This change removes detecting compiled package from the `CompileTask`
and adds a `CompiledPackageFinder` object to consolidate the logic.
Compiled packages are now passed in to the constructor of the
`CompileTask`.

This change also switches the package compile stage to use dependency
injection.

This change also refactors some minor stylistic changes.

[#163832588](https://www.pivotaltracker.com/story/show/163832588)

Co-authored-by: Charles Hansen <chansen@pivotal.io>
Co-authored-by: Danny Berger <dberger@pivotal.io>
Co-authored-by: Belinda Liu <bliu@pivotal.io>