Enhancement for encryption of persistent disk
For redeployment of release with an encrypted disk partition, the file
system is mounted differently. Therefore agent disk cid will be out of
sync with cid stored with director.
This change bypasses that check only if it finds that there is an encryption
package available to enable successful redeployment.
This change will not affect any other scenatio except the one mentioned above.

Co-authored-by: Deepak Shetty <dshetty@zettaset.com>
Co-authored-by: Brian Metzger <bmetzger@zettaset.com>