Enable globally allocated vip network static ips
This commit updates the director to pull from a list of predefined
static ips configured in the cloud config for VIP networks. The
reservation and allocation process is similar to the current process for
manual networks. The director will pick the next available unused static
IP when reserving a VIP for an instance.

If the operator specifies a vip network with static IPs defined on the
instance group, as well as a globally allocated vip network, BOSH will
return a validation error.

Globally allocated VIP networks will prevent instances from being
updated using the `create_swap_delete` strategy.

[#164092238](https://www.pivotaltracker.com/story/show/164092238)

Co-authored-by: Belinda Liu <bliu@pivotal.io>