Avoid sync_dns broadcast to agents during deploy
When a VM is created or deleted as part of a deploy we have previously,
immediately broadcasted sync_dns with new records to all VMs under
director control. In environments with large numbers of VMs and a higher
number of parallel threads, we have observed that the worker processes
start experiencing higher timeouts.

Multiple teams have spent time investigating this issue, but it has been
difficult to pinpoint. One of the theories was that the Ruby worker
process simply couldn't process the queue of NATS messages it emits.

Given that we already have a background job which is sending sync_dns
commands to outdated agents every 10s, we are going to switch to rely on
that broadcast method for regular deploy tasks as well. Historically,
that process was intended as a backup since deploys never waited for
confirmation that VMs received every message happening during the
deploy.

https://github.com/cloudfoundry/bosh/issues/2124
[#163427143](https://www.pivotaltracker.com/story/show/163427143)

Co-authored-by: James Myers <jmyers@pivotal.io>