Allow skipping tests using env param to run_script
Older agents don't accept the env param to run_script, so making
expectations on that feature won't ever work.
This integration test suite will be run with agents that don't have this
feature for compatiblity checking, so we need to have a way to skip
those tests in those runs.

Setting SKIP_RUN_SCRIPT_ENV=true when running integration will now skip
the tests on pre_stop that make use of that feature.