Correctly reuse existing blobs from other release packages
This change updates the behavior of the `update_release` job to
correctly reuse packages from other releases when their fingerprints
match. Previously, the `update_release` would only reuse these packages
if the package had not been created yet, or if the package was already
associated with the existing release version.

This fixes an issue that was seen when uploading source releases to a
director after previously uploading compiled releases. Specifically, if
two releases shared a package with the same fingerprint, when uploading
the second source release, the director would fail to untar the source
package as it was not uploaded. Now the director will reuse the blob
from the first source release.

[#165861952](https://www.pivotaltracker.com/story/show/165861952)

Co-authored-by: Miguel Verissimo <mverissimo@pivotal.io>