Remove filtering by variable types in API
Bosh can't possibly know the variable type when variables are just
referenced in runtime/cloud configs or the manifest, and not defined in the `variables` section.
Providing an unknown type for reference-only variables doesn't provide enough value for downstream users.

The expectation is therefore: the config-server implementation can provide this information via its APIs.

[#165553027](https://www.pivotaltracker.com/story/show/165553027)