Deprecate disk_pools key in manifest
This removes support for any `disk_pools` defined in a deployment
manifest. Given that some disk_pool parsing
occured in the cloud_manifest_parser, it seems that disk_pool
definitions could also appear in the cloud config, but this is
undocumented. If a user now defines a disk_pool in their cloud config,
it will be silently ignored. If a user defines a disk_pool in their
deployment manifest, it will now raise a deprecation error.

[#164777217](https://www.pivotaltracker.com/story/show/164777217)

Co-authored-by: Belinda Liu <bliu@pivotal.io>