Add new endpoint for starting instances independently
This change adds a new endpoint,
`/deployments/:deployment/instance_groups/:instance_group/:index|:uuid/actions/start`
which will start the specified instance independent of the rest of the
deployment. This means that the instance will only run through the
lifecycle methods for starting. It currently does not support starting
instances that have been stopped with the `--hard` option.

This change also refactors the StateApplier to use a new shared class
for starting instances, Starter. This mirrors the existing class for
stopping instances.

This change also renamed the `_skip_post_start` option to be more
correct in that it skips the behavior of waiting for the processes
to successfully start.

[finishes #166769706](https://www.pivotaltracker.com/story/show/166769706)

Co-authored-by: Belinda Liu <bliu@pivotal.io>