Make package_persister into a stateless helper
It was never really a BaseJob; that was just a convenience to get access
to `single_step_stage`. Based on our conversation today I'm duplicating
that method here to make this class into a stateless helper. I was also
able to make every method except those used outside this class private
methods as part of this change.