Remove usages of old postgres/mysql docker files
This commit removes using docker files we no longer want to maintain to run unit
tests. This pipeline is still paused as this director version is at the
end of support, but we should ensure that if we need to run the pipeline
at some point it is able to succeed.

[#168653211](https://www.pivotaltracker.com/story/show/168653211)