Remove state from desired instance
We previously introduced the `state` property in
e31c819cadf481393ca8fd67b765808259f19105 to track the desired state in
instance construction. This code path has since been altered and having
`state` as a property on the class does not provide any value. This
commit removes that property, and also tidies up the various usages of
the DesiredInstance object. It seems that the constructor of the class
has changed over the years and many invocations were stale and
nonsensical.

[#167363054](https://www.pivotaltracker.com/story/show/167363054)