Setup App instance before it is used
The state_applier_spec was not setting up the App instance before
because it was not used. Now, however, we access it in #apply_vm_state
on the DeploymentPlan::Instance, which requires it to be set up since we
are using a real object for that collaborator in the test.

[#168818010](https://www.pivotaltracker.com/story/show/168818010)

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>