Add prometheus metrics for the director api
Prometheus::Collector collects http request metrics.
Since we are using Puma with multiple processes, the metrics need to be
aggregated on the file system, using the DirectFileStore. If the default
Datastore is used, everything will "work", but you will get different
metrics depending on which process responds to the request.

Prometheus::Exporter adds a `/metrics` endpoint which serves the metrics
from the Collector

Local `/metrics` endpoint is hidden and can only be accessed by proxying
through the metrics_server. e.g. "BOSH_IP:9091/director_metrics". The
metrics server is secured with mTLS

- disable metrics collector/exporter if metrics_server is disabled
- update metrics BRAT

[#169331435](https://www.pivotaltracker.com/story/show/169331435)

Co-authored-by: Maya Rosecrance <mrosecrance@pivotal.io>