Stop removing network information from errand VMs
This change updates the errand instance group provider to no longer
remove network information from the instance spec.

This was previously done because the BOSH agent on a GCP VM would return
a netmask that was different than the one that BOSH would calculate.
After saving this different netmask into the instance spec, the director
would then recreate the errand VM when updating the instance, which
resulted in errand VMs being created twice during a `run-errand`.

The changes in https://github.com/cloudfoundry/bosh-google-cpi-release/pull/314
result in VMs being created with the same netmask that the director
calculates for a subnet. Now that version 30 of the google CPI is
released, this change is no longer necessary.

This change also had a side effect in which errand VMs would not receive
a sync dns message synchronously when updating the instance, and would
need to rely on the background sync dns scheduler to eventually send the
message. Now errand VMs will receive the sync dns message in the process
of updating the instance.

[finishes #166359647](https://www.pivotaltracker.com/story/show/166359647)