Remove resetting num tasks to 0
We believe this is redundant, and only present for tests. The prometheus
ruby library API is kinda funky. Ideally, when we fetch a value for a
metric, it would be for the current time. Instead, the library was
fetching the last known data.

In a tool like Grafana, the data for the metrics stops appearing.

[#169881268](https://www.pivotaltracker.com/story/show/169881268)

Co-authored-by: Shaan Sapra <ssapra@pivotal.io>