metrics_server uses "interval" schedule type instead of "every"
This means there are 30 seconds in between each update, instead of
updating every 30 seconds. This will prevent tasks from piling up if it
takes more than 30 seconds to calculate metrics

It is possible for a cloud config with a /16 worth of IP
addresses in reserved or static to take a few minutes.

[#169881268](https://www.pivotaltracker.com/story/show/169881268)

Co-authored-by: Joseph Palermo <jpalermo@pivotal.io>