Use `application/octet-stream` as default type for blobstore nginx
This change updates the `default_type` of the blobstore nginx to be
`application/octet-stream`. Formerly, when the `default_type` was set to
`text/html`, nginx would attempt to compress any assets that were being
fetched from the blobstore. Unfortunately, most of the assets that are
in the blobstore are already compressed making this less performant. By
changing the default to `application/octet-stream` the blobstore will no
longer attempt to compress blobs from the blobstore.

This change also improves the performance of the blobstore in general.
We no longer see TLS handshake timeouts when agents are downloading
multiple assets from the blobstore in parallel.

[#170268644](https://www.pivotaltracker.com/story/show/170268644)
[#170122072](https://www.pivotaltracker.com/story/show/170122072)

Fixes #2225