Allow operators to configure whether the metrics server listens externally
This change adds a
`properties.director.metrics_server.listen_on_external_interface`
property which dictates whether or not the metrics server will be
accessible on the external interface of the director VM. By default,
this value will be `true`. If the property is set to `false`, the
metrics server will only listen on the `backend_port`. Also, the TLS
credentials will not be required.

Fixes #2235