Cloud configs without networks still emit metrics
In the docs we say that cloud configs require networks. Networks can
still be omitted, as when a deployment occurs, the accumulated cloud
configs are then validated. Therefore, it is possible to have cloud
configs without networks.

This change allows for all or some cloud configs to omit networks, and
metrics can still successfully be emitted for network-related metrics.
If no networks exist at all (usually a pointless usecase), then the
metrics will be zero for label 'no_networks'.

Fixes #2237

[#171253546](https://www.pivotaltracker.com/story/show/171253546)