Stop eventmachine loop when error occurs
Exceptions in the event loop of EM are puting the HM in a state,
where the HM server is stopped and also the periodic task,
which raised the exception is canceled. A partially running state
of HM doesn't make sense and it's better to stop the HM.
The HM doesn't have any state, which is lost if it's stopped and
the HM is started again by monit.

fixes #2262