Fix "delivery" job not running
* This is not yet confirmed with the Concourse team
* There appears to be a bug with "version: every" in Concourse 6.0, in
that it was not triggering this job with the newest set of inputs.
* Specifically, a version of "bosh-src" had passed through all three of
the "delivery" jobs passed constraints
(52a80a6bd9be9d16cc8eaa5dd46fc1bdea61f72f), but was not
triggering/passed into the delivery job.
* Temporarily removing this "version: every" fixed the issue.

My suspicion for why it was there, was for the tracker delivery job
below.

With Kevin Rutten