Include orphaned IP addresses in reservation
Previously, the instance network reservation object was only looking at
the IPs that were assigned to the last VM record. This could lead to
situations where, when changing the network of a VM and a deploy failed,
the Director would continuously reserve new IP addresses for the
instance.

Now IP addresses that are associated with the instance but not a VM are
added to the list, allowing them to be used and released properly.

[#177078107] cloudfoundry/bosh #2297: IP address allocations not released on failure while migrating networks of an instance_group with multiple instances

Signed-off-by: Brian Upton <bupton@vmware.com>