Remove legacy stemcell-copy script
We removed an unneeded sudoer privileges and associated scripts. These
used to be used when deploying on AWS. However, this has long since been
moved inside of the AWS CPI. We checked the BOSH Director codebase and
all major CPIs, and AWS was the only one that had references to
`stemcell-copy`. The AWS CPI provides its own version, so the version in
the Director was not used.

Signed-off-by: Brian Upton <bupton@vmware.com>