Fix error messages when BOSH director is down
The Health Monitor endpoint helpers uses EventMachine::HttpClient to
talk to the BOSH Director. This class once had a `uri` method, but it
was removed in later versions of the `em_http_request` gem. So when we
were generating the error messages upon a non-200 Director response,
attempting to include the URI cause an `undefined method` error to be
added to logs instead.

The WebMock gem unfortunately defines the `uri` method on its equivalent
mock class. So our existing tests did not catch this error.

[#177402668] Health monitor should not print stack traces when Director is not up

Signed-off-by: Brian Upton <bupton@vmware.com>