Don't encode query and path in api helper
There is no need to escape input since this code is
only used in our own integration tests so the source can be trusted.

Alternativly we could in the future change query to be a hash and
escape using URI.encode_www_form, but this would mean we have to find
all callers of this method.

[#178224013]

Signed-off-by: Ruben Koster <rubenk@vmware.com>