Fix console test ruby 3.0 compatibility
Ruby 2.7 introduced colors for irb which causes our matchers to fail.
This commit writes a config file to the user home dir before invoking
the console to disable colors.

In addition the interactive prompt uses escaped chars to update itself
which are incompatible with our testing library. To work around this
the irb input is provided via an echo to pipe construct.

Last but not least the pty library has been moved to a new github org
and has been updated.

[#178224013]