Pre-populate blobstore and nats shas in migration
This pre-populates the new blobstore_config_sha1 and nats_config_sha1
columns with the current SHAs when the database is migrated to add the
columns. These values are read off of the Bosh::Director::Config global
class because:

1. The Sequel gem's migration system does not allow outside parameters
   to be passed to the migrations. So it is not possible to pass in the
   full config object directly.
2. The config file path is passed in to the migrate script, so we cannot
   load the file directly in the migration (unless we assume the config
   file path)

Loading the global state onto the Config method required calling the
aptly named `configure_evil_config_singleton!` method, previously only
called when the BOSH Director app is launched.

[#179396629] TAS-29 Rotation of certificates/credentials stored in the bosh-agent metadata can be updated without recreating the VM