Test BATS against Xenial again
The VMware Bosh Ecosystem team has a need to cross-check the behavior of
BATS on both the (VMware-"proprietary") Xenial ESM stemcell, as well as
the latest Open Source stemcell. If you are an Open Source user, and do
not have access to the Xenial ESM stemcell, do know that you can comment
out the `bats-ubuntu-xenial` job and all resources that it references
and have no loss of pipeline functionality.

If this change is inconvenient, or otherwise ruins your day, do accept
my apologies and feel free to send a nastygram to the email address
of the author of this commit.

NOTE: The `bats-ubuntu-xenial` job _should not_ block the rest of the
pipeline if it fails. This is just a cross-check to determine if a BATS
failure is due to a change in the Bionic stemcell, or a change in one of
the components in that stemcell that are managed by the VMware Bosh
Ecosystem team.

This reverts commit 09d56baf728f4f6ccc5ef2d9b42febfe67a336de and adds an
inline comment that explains that the `bats-ubuntu-xenial` job can be
commented out if you don't have access to the Xenial ESM stemcells.