Fixes TarGzipper unit spec
This commit makes the affected test stop using the `base_dir` variable
to pass an argument to the `TarGzipper::compress` function. The `base_dir`
variable is used elsewhere in the test suite setup code, so setting this
to invalid values has unexpected side effects.

Specifically, the `before` block for the tests in this file create the
directory specified in the `base_dir` variable.

Without this change, the `before` block attempts to create a deep directory
that doesn't exist, raises an exception,  and the affected test fails with
the following message:

========================
Failures:

  1) Bosh::Director::Core::TarGzipper if the source directory does not exist raises an error
     Failure/Error: FileUtils.mkdir_p(path)

     Errno::EACCES:
       Permission denied @ dir_s_mkdir - /tmp/this/is/not/here
     # ./spec/bosh/director/core/tar_gzipper_spec.rb:19:in `block (2 levels) in <module:Core>'

Finished in 0.17048 seconds (files took 0.59217 seconds to load)
6 examples, 1 failure

Failed examples:

rspec ./spec/bosh/director/core/tar_gzipper_spec.rb:56 # Bosh::Director::Core::TarGzipper if the source directory does not exist raises an error
========================

Given that the intent of the test seems to be to test the behavior of the
`TarGzipper::compress` function, rather than the standard library's directory
creation code, this seems like a reasonable change to make. Given that there's
no reason to remove a directory that doesn't exist, the call to
`FileUtils.rm_rf` is also removed by this commit. (I suspect that it was
added in as a way to side-step the side-effect of the before block
actually creating the path in question.)