Adds test-running and minimum Ruby version docs
This commit tells folks the minimum version of Ruby required to run the
tests (and -presumably- run the software under test), as well as how to
run the unit tests.

Because neither the CONTRIBUTING file, nor the README file currently
contain this information, I'm putting it in the README file.